move log to utils
