+++ app/domain/shipper/shipper.go
@@ -17,10 +17,10 @@ import (
 	"time"
 
 	config "github.com/cloudzero/cloudzero-agent/app/config/gator"
-	"github.com/cloudzero/cloudzero-agent/app/lock"
 	"github.com/cloudzero/cloudzero-agent/app/logging/instr"
 	"github.com/cloudzero/cloudzero-agent/app/storage/disk"
 	"github.com/cloudzero/cloudzero-agent/app/types"
+	"github.com/cloudzero/cloudzero-agent/app/utils/lock"
 	"github.com/cloudzero/cloudzero-agent/app/utils/parallel"
 	"github.com/rs/zerolog"
 	"github.com/rs/zerolog/log"
